[FIX] fix GIL testing
- add an easy way to check for the GIL in Python 3.4 and higher
- fix issue #36

it seems that (PyThreadState*)_Py_atomic_load_relaxed(&_PyThreadState_Current);
does not work any more for Python 3.5 as the atomic functions are only
available for C++11 and higher using the atomic.h package. This is a
known bug in Python, see https://bugs.python.org/issue23644 as #include
<stdatomic.h> fails in C++. So therefore, from Python 3.5 on we cannot
use this any more to check for the GIL. Fortunately there is an easier
way to perform the same check.